requirements.txt
Flask>=2.2
flask-cors
python-dotenv
PyMuPDF>=1.22.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
numpy
requests
gunicorn
pytest
pytest-mock
